cleaning and refactoring
